Move modular objects into shared_utils
